PIM-5210: Respect current mass edit common attributes behavior
