+++ features/Context/Page/Base/ProductEditForm.php
@@ -5,6 +5,7 @@ namespace Context\Page\Base;
 use Behat\Mink\Element\Element;
 use Behat\Mink\Element\NodeElement;
 use Behat\Mink\Exception\ElementNotFoundException;
+use Behat\Mink\Exception\ExpectationException;
 
 /**
  * Product Edit Form
@@ -558,5 +559,185 @@ class ProductEditForm extends Form
                 )
             );
         });
+    }
+
+    /**
+     * Checks if the specified field is set to the expected value, raises an exception if not
+     *
+     * @param string $label
+     * @param string $expected
+     * @param bool   $copy
+     *
+     * @throws ExpectationException
+     */
+    public function compareFieldValue($label, $expected, $copy = false)
+    {
+        $fieldContainer = $this->findFieldContainer($label);
+        $fieldType      = $this->getFieldType($fieldContainer);
+        $subContainer   = $fieldContainer->find('css', $copy ? '.copy-container .form-field' : '.form-field');
+
+        switch ($fieldType) {
+            case 'textArea':
+                $actual = $this->getTextAreaFieldValue($subContainer);
+                break;
+            case 'metric':
+                $actual = $this->getMetricFieldValue($subContainer);
+                break;
+            case 'multiSelect':
+                $actual   = $this->getMultiSelectFieldValue($subContainer);
+                $expected = $this->listToArray($expected);
+                sort($actual);
+                sort($expected);
+                $actual   = implode(', ', $actual);
+                $expected = implode(', ', $expected);
+                break;
+            case 'select':
+                $actual = $this->getSelectFieldValue($subContainer);
+                break;
+            case 'media':
+                $actual = $this->getMediaFieldValue($subContainer);
+                break;
+            case 'switch':
+                $actual   = $this->isSwitchFieldChecked($subContainer);
+                $expected = ('on' === $expected);
+                break;
+            case 'text':
+            case 'date':
+            case 'number':
+            case 'price':
+            default:
+                $actual = $this->findField($label, $copy)->getValue();
+                break;
+        }
+
+        if ($expected != $actual) {
+            throw new ExpectationException(
+                sprintf(
+                    'Expected product field "%s" to contain "%s", but got "%s".',
+                    $label,
+                    $expected,
+                    $actual
+                ),
+                $this->getSession()
+            );
+        }
+    }
+
+    /**
+     * Returns the current value of a textarea
+     * Handles both simple textarea and wysiwyg editor
+     *
+     * @param NodeElement $subContainer
+     *
+     * @return string
+     */
+    protected function getTextAreaFieldValue(NodeElement $subContainer)
+    {
+        $field = $subContainer->find('css', '.field-input textarea');
+
+        if (!$field || !$field->isVisible()) {
+            // the textarea can be hidden (display=none) when using WYSIWYG
+            $div = $subContainer->find('css', '.note-editor > .note-editable');
+
+            return $div->getHtml();
+        } else {
+            return $field->getValue();
+        }
+    }
+
+    /**
+     * Return the current formatted value of a metric field (e.g.: '4 KILOGRAM')
+     *
+     * @param NodeElement $subContainer
+     *
+     * @return string
+     */
+    protected function getMetricFieldValue(NodeElement $subContainer)
+    {
+        $input  = $subContainer->find('css', '.data');
+        $select = $this->spin(function () use ($subContainer) {
+            return $subContainer->find('css', '.select2-container');
+        }, 5);
+
+        return sprintf(
+            '%s %s',
+            $input->getValue(),
+            $select->find('css', '.select2-chosen')->getText()
+        );
+    }
+
+    /**
+     * Return the current values of a multi-select field
+     *
+     * @param NodeElement $subContainer
+     *
+     * @return array
+     */
+    protected function getMultiselectFieldValue(NodeElement $subContainer)
+    {
+        $input = $this->spin(function () use ($subContainer) {
+            return $subContainer->find('css', 'input[type="hidden"].select-field');
+        }, 5);
+
+        return '' === $input->getValue() ? [] : explode(',', $input->getValue());
+    }
+
+    /**
+     * Return the current value of a select field
+     *
+     * @param NodeElement $subContainer
+     *
+     * @return string
+     */
+    protected function getSelectFieldValue(NodeElement $subContainer)
+    {
+        $input = $this->spin(function () use ($subContainer) {
+            return $subContainer->find('css', 'input[type="hidden"].select-field');
+        }, 5);
+
+        return $input->getValue();
+    }
+
+    /**
+     * Return the current filename uploaded in a media field
+     *
+     * @param NodeElement $subContainer
+     *
+     * @return string
+     */
+    protected function getMediaFieldValue(NodeElement $subContainer)
+    {
+        $widget = $this->spin(function () use ($subContainer) {
+            return $subContainer->find('css', '.field-input .media-uploader');
+        }, 5);
+
+        $filenameNode = $widget->find('css', '.filename');
+
+        return $filenameNode ? $filenameNode->getText() : '';
+    }
+
+    /**
+     * Return the state of a switch field
+     *
+     * @param NodeElement $fieldContainer
+     *
+     * @throws \LogicException
+     *
+     * @return bool
+     */
+    protected function isSwitchFieldChecked(NodeElement $fieldContainer)
+    {
+        $widget = $this->spin(function () use ($fieldContainer) {
+            return $fieldContainer->find('css', '.field-input .switch.has-switch');
+        }, 5);
+
+        if ($widget->find('css', '.switch-on')) {
+            return true;
+        }
+        if ($widget->find('css', '.switch-off')) {
+            return false;
+        }
+
+        throw new \LogicException(sprintf('Switch "%s" is in an undefined state', $fieldContainer->name));
     }
 }
