+++ spec/Pim/Bundle/EnrichBundle/Connector/Processor/MassEdit/Product/EditCommonAttributesProcessorSpec.php
@@ -87,7 +87,8 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
                 [
                     'filters' => [],
                     'actions' => [
-                        'normalized_values' => $normalizedValues
+                        'normalized_values' => $normalizedValues,
+                        'current_locale'    => 'en_US'
                     ]
                 ]
             )
@@ -135,7 +136,8 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
                 [
                     'filters' => [],
                     'actions' => [
-                        'normalized_values' => $normalizedValues
+                        'normalized_values' => $normalizedValues,
+                        'current_locale'    => 'en_US'
                     ]
                 ]
             )
@@ -183,7 +185,8 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
                 [
                     'filters' => [],
                     'actions' => [
-                        'normalized_values' => $normalizedValues
+                        'normalized_values' => $normalizedValues,
+                        'current_locale'    => 'en_US'
                     ]
                 ]
             )
