+++ src/Pim/Bundle/EnrichBundle/Form/Type/MassEditAction/EditCommonAttributesType.php
@@ -59,6 +59,7 @@ class EditCommonAttributesType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder->add('values', 'hidden');
+        $builder->add('current_locale', 'hidden');
     }
 
     /**
