+++ src/Pim/Bundle/EnrichBundle/Resources/views/MassEditAction/configure/edit-common-attributes.html.twig
@@ -6,6 +6,7 @@
     </div>
 
     {{ form_widget(form.operation.values) }}
+    {{ form_widget(form.operation.current_locale) }}
 
     <script>
         require(
